# cmd/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the ArchiD-Projet project",
        "title": "ArchiD-Projet API",
        "contact": {},
        "version": "1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/airport/{iata}/average/{date}": {
            "get": {
                "description": "Get average data for an airport for a date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get average data for an airport for a date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airport IATA code",
                        "name": "iata",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.dataAverage"
                            }
                        }
                    }
                }
            }
        },
        "/airport/{iata}/average/{date}/{measurement}": {
            "get": {
                "description": "Get average data for an airport for a date and a measurement type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get average data for an airport for a date and a measurement type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airport IATA code",
                        "name": "iata",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Measurement type",
                        "name": "measurement",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.dataAverage"
                            }
                        }
                    }
                }
            }
        },
        "/airport/{iata}/data": {
            "get": {
                "description": "Get all data for an airport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all data for an airport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airport IATA code",
                        "name": "iata",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.data"
                            }
                        }
                    }
                }
            }
        },
        "/airport/{iata}/data/range/{start}/{end}/{measurement}": {
            "get": {
                "description": "Get all data for an airport between two dates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all data for an airport between two dates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airport IATA code",
                        "name": "iata",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date",
                        "name": "start",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date",
                        "name": "end",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Measurement type",
                        "name": "measurement",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.data"
                            }
                        }
                    }
                }
            }
        },
        "/airport/{iata}/sensors": {
            "get": {
                "description": "Get all sensors for an airport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all sensors for an airport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airport IATA code",
                        "name": "iata",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.sensor"
                            }
                        }
                    }
                }
            }
        },
        "/airports": {
            "get": {
                "description": "Get all airports",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all airports",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.airport"
                            }
                        }
                    }
                }
            }
        },
        "/airports/data": {
            "get": {
                "description": "Get all data for all airports",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all data for all airports",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.data"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.airport": {
            "type": "object",
            "properties": {
                "airport": {
                    "type": "string"
                }
            }
        },
        "main.data": {
            "type": "object",
            "properties": {
                "airport": {
                    "type": "string"
                },
                "measurement": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "main.dataAverage": {
            "type": "object",
            "properties": {
                "airport": {
                    "type": "string"
                },
                "measurement": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "main.sensor": {
            "type": "object",
            "properties": {
                "airport": {
                    "type": "string"
                },
                "measurement": {
                    "type": "string"
                }
            }
        }
    }
}